// frontend/src/app/api/model/publisherDto.ts
/**
 * Book Shelf API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PublisherDto {
  /**
   * Unique publisher identifier.
   */
    id?: string;
  /**
   * Publisher name.
   */
    name?: string;
}